==> 1_openai_gather.ipynb <==
# Imports

import whisper
import pandas as pd

model = whisper.load_model("base")

# Transcrever episódios

start = 385
stop = 405
for num in range(start, stop+1):
    filename = f"xadrez_verbal_{num}.mp3"

    print(f"### === Começando transcrição do programa Nº {num} === ###")
    result = model.transcribe("Programas/"+filename, language="pt", fp16=False)

    print(f"### === Programa {num} transcrito! Iniciando download === ###")
    with open("TextFiles/"+filename[:-3]+"txt", "w", encoding='utf-8') as textfile:
        textfile.write(result['text'])

    dados = []
    for segment in result["segments"]:
        dados.append({
            "start_segment": round(segment["start"], 2),
            "end_segment": round(segment["end"], 2),
            "text": segment["text"].strip()
        })

    df = pd.DataFrame(dados)

    df.to_csv("Tables/"+filename[:-3]+"csv", index=False, encoding="utf-8-sig")

    print(f"### === Processo concluído para o Programa {num} === ###")
