==> 4_create_tables.ipynb <==
# Imports

import pandas as pd 

# Definir função que salva tabela pandas como código HTML formatado

def table_to_html(df, save_location, add_text=""):
    # Generate the basic table HTML
    html_table = df.to_html(classes="table table-sm table-bordered", index=False, border=0)

    # Wrap it with full HTML, including Bootstrap CSS
    full_html = f"""
    <!DOCTYPE html>
    <html lang="en">
    <head>
        <meta charset="UTF-8">
        <title>Table</title>
        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet">
        <style>
            table {{
                font-size: 11px;
                font-family: Verdana;
                
            }}
            table th {{
                text-align: center;
                vertical-align: middle;
                }}
        </style>
    </head>
    <body>
    <div class="container my-4">
        {html_table}
        <p style="font-size:10px">{add_text}</p>
    </div>
    
    </body>
    </html>
    """

    # Save to file
    with open(save_location, "w", encoding="utf-8") as f:
        f.write(full_html)


changecols = {'Pais':'País', 'NumCit': 'Número de Citações', 'NumProgramas':'Número de Programas'}

## Países Mais Citados

# Paises mais citados
df_cit = pd.read_csv('Data/Numero_Citacoes.csv')
df_prog = pd.read_csv('Data/Numero_Programas.csv')
df_tot = df_cit.merge(df_prog, on='Pais', how='outer')

df_mais_citados = df_tot.sort_values(by='NumCit', ascending=False).head(10).rename(columns=changecols).reset_index(drop=True)
df_mais_citados

table_to_html(df_mais_citados, "Pages/table_mais_citados.htm")

## Países citados em mais programas

df_mais_programas = df_tot[df_tot['NumProgramas']==max(df_tot['NumProgramas'])].sort_values(by='NumCit', ascending=False).rename(columns=changecols).reset_index(drop=True)
df_mais_programas


table_to_html(df_mais_programas, "Pages/table_mais_programas.htm")

## Países com maior relação entre si

df_rel = pd.read_csv('Data/Relacoes_Paises_Contagens.csv')
df_rel_show = df_rel.sort_values(by='num_relac_paises', ascending=False).head(10)[['pais1', 'pais2', 'num_relac_paises']].rename(columns={'pais1': 'País 1', 'pais2': 'País 2', 'num_relac_paises': 'Citações Conjuntas'})

table_to_html(df_rel_show, "Pages/paises_mais_relacao.htm", add_text="* Países citados conjuntamente são países citados na mesma frase.")

## Países mais relacionados com o Brasil

df_brasil_rel_show = df_rel[(df_rel['pais1']=='Brasil') | (df_rel['pais2']=='Brasil')].sort_values(by='num_relac_paises', ascending=False).head(10)[['pais1', 'pais2', 'num_relac_paises']].rename(columns={'pais1': 'País 1', 'pais2': 'País 2', 'num_relac_paises': 'Citações Conjuntas'})
df_brasil_rel_show

table_to_html(df_brasil_rel_show, "Pages/brasil_mais_relacao.htm", add_text="* Países citados conjuntamente são países citados na mesma frase.")
